[UPDATE] 1. Try Fix FtpSetting account and password is null problem

+++ com.bcs.core.taishin/src/main/java/com/bcs/core/taishin/circle/db/repository/BillingNoticeDetailRepository.java
@@ -27,6 +27,6 @@ public interface BillingNoticeDetailRepository extends EntityRepository<BillingN
     @Transactional(rollbackFor = Exception.class, timeout = 30)
     void updateStatusByMainId(String status, Date modifyTime, Long noticeMainId);
 
-    List<BillingNoticeDetail> findByNoticeDetailIdIn(List<BigInteger> ids);
+    List<BillingNoticeDetail> findByNoticeDetailIdIn(List<Long> ids);
 
 }
